Workflow to Provision/Deprovision a full TM/DB IMS System
This will provision a full IMS TM/DB system starting with the SMP
installed IMS datasets that must already exist.  The Provision workflow
will not modify the datasets, instead making copies in a temporary
location of datasets that need modification.  This allows the
Deprovision workflow to take down the IMS regions and delete the
temporary datasets, thus leaving the target system in the exact state
that exists prior to running the Provision workflow.