Provision/Deprovision a FastPath database
This will provision a FastPath database onto an existing IMS.  The
structure of the database is fixed, but names can be modified at
runtime.  The Provision workflow can be run multiple times to create as
many databases as the user chooses.  The Deprovision workflow will
delete a selected database and leave all others intact.